requirements.txt
lz4
Pillow
texture2ddecoder
